consolidate remaining old table assets under datamap
