+++ clients/admin-ui/src/features/datamap/datamap-table/react-table-config.d.ts
